Fix character movement
Fix character movement as sometimes there was a rounding float error
which was causing to select the wrong cell. Make sure we always select
the correct cell now. Also add button to clear all collision as well as
made the 2 cubes have the same color of their correspectivec paths to be
clearer which is which